Moves DriverStation HAL code to its own header and definition
To allow for future use as a plugin for the HAL simulator.
